refactor WIP Issue: AAH-2125
