+++ galaxy_ng/tests/integration/api/test_custom_repo_sync.py
@@ -6,11 +6,8 @@ import pytest
 from galaxykit.remotes import create_remote
 from galaxykit.repositories import create_repository
 from galaxykit.utils import wait_for_task
-from .test_x_repo_search import create_test_namespace, upload_new_artifact, search_collection_endpoint, \
-    add_content_units, create_repo_and_dist, verify_repo_data
-from .test_sync_stage import start_sync
-from ..conftest import AnsibleConfigFixture, get_galaxy_client, get_ansible_config
-from ..utils import get_client, clear_certified, perform_sync, iterate_all, set_synclist
+from ..utils.repo_management_utils import create_repo_and_dist, search_collection_endpoint, create_test_namespace, \
+    upload_new_artifact, add_content_units, verify_repo_data
 from ..utils.iqe_utils import GalaxyKitClient
 from ..utils.tools import generate_random_string
